OBD2: service-aware (svc,pid) tuples + add Toyota ATF temp PID
Phase 1: refactor the OBD2 layer so a PID byte can mean different
things in different services without ambiguity. Phase 2: add the
first Toyota Mode 21 PID that needed this — ATF temp (PID 0x21),
which would have collided with Mode 01 PID 0x21 (DTC distance) under
the old byte-only lookup.

Encoding scheme:
   value = (service << 8) | pid
 - Values <= 0xFF treated as Mode 01 (backwards compat for layouts
   written by earlier firmware that stored bare PID bytes).
 - obd2_encode_pid() / obd2_decode_service() / obd2_decode_pid()
   helpers in obd2.h keep the encoding ergonomic.

Public API change (uint8_t -> uint16_t):
 - obd2_start(const uint16_t *enc, count)
 - obd2_get_enabled(uint16_t *out, max) -> returns encoded tuples
 - obd2_set_enabled(const uint16_t *enc, count)

Service-aware lookup:
 - New obd2_pid_find_svc(service, pid) is the canonical lookup.
   Falls back to first-match-by-byte when no exact match exists
   (preserves behaviour for callers that don't know the service).
 - Internal _find_for_service helper deleted; all sites now use the
   public obd2_pid_find_svc directly.
 - obd2_pid_find(pid) kept for live-indicator polling (just needs
   any match by byte for tier hints).

JSON storage:
 - polled_pids array now emits encoded numbers. Back-compat read:
   numbers <= 0xFF mean Mode 01, larger numbers carry the encoded
   tuple. Legacy `obd2_pids` key still accepted as a fallback.

preconfig_item_t:
 - New `obd2_service` field so the long-press preset picker tracks
   which service each row's PID belongs to. _add_obd2_item populates
   it from the def. _apply_obd2_preset in config_modal uses
   (service, pid) to enable the right entry — Toyota Mode 21 picks
   no longer accidentally trigger Mode 01 lookups.

obd2_picker:
 - Snapshot (s_snapshot) and preview-poll-all list now uint16_t
   encoded.
 - Row enabled-state matching uses encoded tuples instead of bare
   PID bytes.
 - Scan-complete supports only Mode 01 PIDs (scan probes Mode 01
   bitmasks); row match restricted to parent_service == 0x01.
 - Save collects distinct (service, pid) tuples from checked rows.

Toyota Extended pack — first new member:
 - PID 0x21 (Mode 21): Toyota ATF temperature. Single-value, byte 0
   = temp - 40 °C. Addressed to TCM (0x7E1 -> 0x7E9). Coexists with
   Mode 01 PID 0x21 (DTC distance) cleanly now that lookups are
   service-aware. Category="Toyota" so it groups under
   "Toyota / Mode 21" in the long-press picker.
 - More Toyota PIDs deferred — adding speculative decodes without
   empirical testing on real cars produces noise. The architectural
   work here unlocks adding any (service, pid) tuple safely, so the
   pack grows as users test and report what works.

+++ main/layout/ecu_presets.h
@@ -95,30 +95,32 @@ esp_err_t ecu_preset_apply_to_layout(const char *layout_name,
 bool ecu_preset_is_obd2(const ecu_preset_t *preset);
 
 /**
- * Replace the layout's `obd2_pids` array with the given list. Used by the
+ * Replace the layout's `polled_pids` array with the given list. Used by the
  * OBD2 Signals picker after the user changes the enabled-PID set. Does NOT
  * touch the signals[] array, ecu/ecu_version fields, or widgets.
  *
  * @param layout_name  Layout to rewrite (typically "default").
- * @param pids         PID bytes to enable.
+ * @param pids         Encoded (service, pid) tuples — obd2_encode_pid().
  * @param count        Number of PIDs.
  * @return ESP_OK on success.
  */
 esp_err_t ecu_preset_save_obd2_pids(const char *layout_name,
-                                    const uint8_t *pids, uint8_t count);
+                                    const uint16_t *pids, uint8_t count);
 
 /**
- * Read the layout's `obd2_pids` array. Used by dashboard_init after a layout
- * load to know which PIDs to start polling.
+ * Read the layout's `polled_pids` array. Used by dashboard_init after a
+ * layout load to know which PIDs to start polling. Falls back to the
+ * legacy `obd2_pids` key for older layouts (bare bytes interpreted as
+ * Mode 01).
  *
  * @param layout_name  Layout to read from.
- * @param out          Buffer to receive PID bytes.
+ * @param out          Buffer to receive encoded (service, pid) tuples.
  * @param max          Capacity of @p out.
  * @param out_count    On return, number of PIDs written.
  * @return ESP_OK on success (may return zero PIDs).
  */
 esp_err_t ecu_preset_read_obd2_pids(const char *layout_name,
-                                    uint8_t *out, uint8_t max,
+                                    uint16_t *out, uint8_t max,
                                     uint8_t *out_count);
 
 #ifdef __cplusplus
